Create a new OVS manager for metadata agent
When metadata agent starts, it can't connect to the UNIX socket
of the ovsdb-server because of lack of permissions. Socket is 750
and belongs to 'openvswitch' user while agent is started under
'neutron' user. Hence, this patch makes the default connection via
TCP and, for that, we'll create a manager if it's not there from
puppet.

Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>